+++ packages/reactabular-utils/__tests__/resolve_row_key.js
@@ -1,0 +1,48 @@
+import { expect } from 'chai';
+import { resolveRowKey } from '../src';
+
+describe('resolveRowKey', function () {
+  it('resolves to rowKey', function () {
+    const rowKey = 'id';
+    const rowData = {
+      [rowKey]: 5,
+      foo: 'bar'
+    };
+    const rowIndex = 0;
+
+    expect(resolveRowKey({ rowData, rowIndex, rowKey })).to.deep.equal('5-row');
+  });
+
+  it('resolves to rowIndex if rowKey is not found', function () {
+    const rowKey = 'zoo';
+    const rowData = {
+      id: 5,
+      foo: 'bar'
+    };
+    const rowIndex = 0;
+
+    expect(resolveRowKey({ rowData, rowIndex, rowKey })).to.deep.equal('0-row');
+  });
+
+  it('resolves zero index to zero', function () {
+    const rowKey = 'id';
+    const rowData = {
+      [rowKey]: 0,
+      foo: 'bar'
+    };
+    const rowIndex = 0;
+
+    expect(resolveRowKey({ rowData, rowIndex, rowKey })).to.deep.equal('0-row');
+  });
+
+  it('allows rowKey to be a function', function () {
+    const rowKey = ({ rowIndex }) => rowIndex;
+    const rowData = {
+      [rowKey]: 5,
+      foo: 'bar'
+    };
+    const rowIndex = 0;
+
+    expect(resolveRowKey({ rowData, rowIndex, rowKey })).to.deep.equal(`${rowIndex}-row`);
+  });
+});