+++ packages/reactabular-utils/CHANGELOG.md
@@ -1,3 +1,8 @@
+6.1.3 / 2016-10-31
+==================
+
+  * Bug fix - If `rowKey` value is zero, `resolveRowKey` should return `0-row`.
+
 3.0.0 / 2016-09-01
 ==================